Added Obfuscation and Made Checksumming minimized code easier
A new visitor called ObfuscatingVisitor exists whose purpose is to obfuscate variable names.  Additionally a when --checksum and --minimize are used together the checksum of the original file and the minimized code are both displayed.  This allows developers to ensure that the minimize command is producing correct output.  When --checksum and --minimize are specified together obfuscation is turned off since the result of comparing the checsksum of non-obfuscated code against the checksum of its obfucated counterpart will produce non-meaningful results.