MinimizeCommand now flushes its writer
MinimizeCommand now flushes the writer it uses to write messages to its OutputStream.  Prior to this fix the any content left in the writer's buffer would not be written to the OutputStream before the application closes, mainly the checksum of the minimized source.  Additionally, a flush has been added before parsing the source to give the user an indication that the program is running.