Obfuscated identifiers are now obfuscated using a Base64-like encoding algorithm
A new algorithm for generating obfuscated identifiers has been created.  This algorithm is inspired by the Base64 algorithm with a twist to make it work within the confines of what constitutes a valid identifier in JavaScript.  The new algorithm packs the same information previously used into fewer characters and therefore aids in minimizing.