Added (limited) support for using keywords as identifiers.
jQuery does not adhere completely to the JavaScript specification in that it uses resserved words as keys in an object literal.  Additionally, these are used as the RHS of a member access.  The offending code is as follows:

var x = { is: function() {... } }; // FAIL because 'is' is a reserved word
x.is(); // FAIL because 'is' is a reserved word

Additionally, the parser has been changed to support unicode input, although unicode characters are not recognised as valid characters in any language construct other than comments.